SummaryGroup.js is more generic, moved summaries query to LandingPage.js
