+++ frontend/src/OrganizationSummary.js
@@ -1,11 +1,9 @@
 import React from 'react'
-import { Trans, t } from '@lingui/macro'
+import { Trans } from '@lingui/macro'
 import { Layout } from './Layout'
-import { Text, Stack, Box, SimpleGrid } from '@chakra-ui/core'
-// import { SummaryGroup } from './SummaryGroup'
+import { Text, Stack, Box } from '@chakra-ui/core'
+import { SummaryGroup } from './SummaryGroup'
 import { number, object, string } from 'prop-types'
-import SummaryCard from './SummaryCard'
-import theme from './theme/canada'
 
 export function OrganizationSummary({
   summaries,
@@ -14,7 +12,6 @@ export function OrganizationSummary({
   city,
   province,
 }) {
-  const { colors } = theme
   return (
     <Layout>
       <Box fontSize="xl">
@@ -41,45 +38,7 @@ export function OrganizationSummary({
           </Text>
         </Stack>
       </Box>
-      <SimpleGrid
-        columns={[1, 1, 1, 2]}
-        spacing="30px"
-        justifyItems="center"
-        maxWidth="width.60"
-        mx="auto"
-        p={['2', '8']}
-      >
-        <SummaryCard
-          title={t`Web Configuration`}
-          description={t`Web encryption settings summary`}
-          categoryDisplay={{
-            fail: {
-              name: t`Non-compliant TLS`,
-              color: colors.weak,
-            },
-            pass: {
-              name: t`Policy compliant TLS`,
-              color: colors.strong,
-            },
-          }}
-          data={summaries.web}
-        />
-        <SummaryCard
-          title={t`Email Configuration`}
-          description={t`Email security settings summary`}
-          categoryDisplay={{
-            pass: {
-              name: t`Dmarc pass`,
-              color: colors.strong,
-            },
-            fail: {
-              name: t`Dmarc fail`,
-              color: colors.weak,
-            },
-          }}
-          data={summaries.mail}
-        />
-      </SimpleGrid>
+      <SummaryGroup web={summaries.web} mail={summaries.mail} />
       <Trans>*All data represented is mocked for demonstration purposes</Trans>
     </Layout>
   )
